power of

+++ 0x08-recursion/4-pow_recursion.c
@@ -13,7 +13,7 @@ int _pow_recursion(int x, int y)
 	if (y < 0)
 		return (-1);
 
-	printf("%f", pow(x, y));
+	printf("%f", pow((double) x, (double) y));
 
 	return (0);
 }
